requirements.txt
a-unet==0.0.16
absl-py==2.1.0
accelerate==1.1.1
aiohappyeyeballs==2.4.3
aiohttp==3.11.8
aiosignal==1.3.1
alembic==1.14.0
antlr4-python3-runtime==4.9.3
argbind==0.3.9
asteroid-filterbanks==0.4.0
asttokens==2.4.1
async-timeout==5.0.1
attrs==24.2.0
audioread==3.0.1
babel==2.16.0
bibtexparser==2.0.0b7
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
clldutils==3.24.0
colorama==0.4.6
colorlog==6.9.0
contourpy==1.3.1
csvw==3.5.1
cycler==0.12.1
data-science-types==0.2.23
decorator==5.1.1
descript-audiotools @ git+https://github.com/descriptinc/audiotools@348ebf2034ce24e2a91a553e3171cb00c0c71678
dlinfo==1.2.1
docopt==0.6.2
docstring_parser==0.16
einops==0.8.0
einops-exts==0.0.4
exceptiongroup==1.2.2
executing==2.1.0
ffmpy==0.4.0
filelock==3.16.1
fire==0.7.0
flatten-dict==0.4.2
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.10.0
future==1.0.0
greenlet==3.1.1
grpcio==1.68.0
huggingface-hub==0.26.3
HyperPyYAML==1.2.2
idna==3.10
importlib_resources==6.4.5
ipython==8.29.0
isodate==0.7.2
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
julius==0.2.7
kiwisolver==1.4.7
language-tags==1.2.0
lazy_loader==0.4
librosa==0.10.2.post1
lightning==2.4.0
lightning-utilities==0.11.9
llvmlite==0.43.0
lxml==5.3.0
Mako==1.3.6
Markdown==3.7
markdown-it-py==3.0.0
markdown2==2.5.1
MarkupSafe==3.0.2
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
monotonic_align @ git+https://github.com/resemble-ai/monotonic_align.git@78b985be210a03d08bc3acc01c4df0442105366f
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
munch==4.0.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
optuna==4.1.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
phonemizer==3.3.0
pillow==11.0.0
platformdirs==4.3.6
pooch==1.8.2
primePy==1.3
prompt_toolkit==3.0.48
propcache==0.2.0
protobuf==4.25.5
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.0
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pycparser==2.22
Pygments==2.18.0
pylatexenc==2.10
pyloudnorm==0.1.1
pyparsing==3.2.0
pystoi==0.4.1
python-dateutil==2.9.0.post0
pytorch-lightning==2.4.0
pytorch-metric-learning==2.7.0
pytz==2024.2
PyYAML==6.0.2
randomname==0.2.1
rdflib==7.1.1
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rfc3986==1.5.0
rich==13.9.4
rpds-py==0.21.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
segments==2.2.1
semver==3.0.2
sentencepiece==0.2.0
shellingham==1.5.4
six==1.16.0
sortedcontainers==2.4.0
soundfile==0.12.1
soxr==0.5.0.post1
speechbrain==1.0.2
SQLAlchemy==2.0.36
stack-data==0.6.3
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
termcolor==2.5.0
threadpoolctl==3.5.0
tokenizers==0.20.3
torch==2.5.1
torch-audiomentations==0.11.1
torch-stoi==0.2.3
torch_pitch_shift==1.2.5
torchaudio==2.5.1
torchmetrics==1.6.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.46.3
triton==3.1.0
typer==0.13.1
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
wcwidth==0.2.13
Werkzeug==3.1.3
yarl==1.18.0
